src/backend/requirements.txt
Django>=3.2.0
django-cors-headers
selenium
django-crontab
msedge-selenium-tools
APScheduler
lxml
